pyproject.toml
[project]
name = "ThalamOS"
version = "0.2.3"
description = "Light up your storage"
authors = [{ name = "PixelWelt" }]
requires-python = ">=3.10,<=3.14"
readme = "README.md"
license = "MIT"
keywords = [
    "Storage",
    "API",
    "ThalamOS",
    "WLED",
    "Flask",
    "Web",
    "Python",
    "self-hosted",
    "web",
    "cloud",
    "management-system",
]
dependencies = [
    "aiohappyeyeballs",
    "aiohttp",
    "aiosignal",
    "annotated-types",
    "anyio",
    "attrs",
    "backoff",
    "blinker",
    "cachecontrol",
    "certifi",
    "charset-normalizer",
    "cleo",
    "click",
    "colorama",
    "construct",
    "crashtest",
    "dataclasses-json",
    "distlib",
    "distro",
    "dulwich",
    "fastjsonschema",
    "filelock",
    "Flask",
    "flask-cors",
    "frozenlist",
    "greenlet",
    "gunicorn",
    "h11",
    "haystack-ai",
    "haystack-experimental",
    "httpcore",
    "httpx",
    "httpx-sse",
    "idna",
    "iniconfig",
    "itsdangerous",
    "Jinja2",
    "jiter",
    "jsonpatch",
    "jsonpointer",
    "jsonschema",
    "jsonschema-specifications",
    "keyring",
    "lazy_imports",
    "loguru",
    "MarkupSafe",
    "marshmallow",
    "monotonic",
    "more-itertools",
    "msgpack",
    "multidict",
    "mypy-extensions",
    "networkx",
    "numpy",
    "ollama",
    "ollama-haystack",
    "openai",
    "orjson",
    "packaging",
    "pandas",
    "pandas-stubs~=2.3.3",
    "pefile",
    "pkginfo",
    "platformdirs",
    "pluggy",
    "posthog",
    "propcache",
    "pydantic",
    "pydantic-settings",
    "pydantic_core",
    "pytest",
    "python-dateutil",
    "python-dotenv",
    "python-ptrace",
    "pytz",
    "pywin32-ctypes",
    "PyYAML",
    "RapidFuzz",
    "referencing",
    "requests",
    "requests-toolbelt",
    "rpds-py",
    "shellingham",
    "six",
    "sniffio",
    "SQLAlchemy",
    "sqlmodel>=0.0.33",
    "sqlparse",
    "tenacity",
    "tomlkit",
    "tqdm",
    "trove-classifiers",
    "typing-inspect",
    "typing_extensions",
    "tzdata",
    "urllib3",
    "virtualenv",
    "Werkzeug",
    "win32_setctime",
    "yarl",
    "zstandard",
]

[project.urls]
Repository = "https://github.com/PixelWelt/ThalamOS"
Documentation = "https://pixelwelt.github.io/ThalamOS/"
